Magic Matrices Task
Function that checks if a given matrix of numbers is magical. A matrix is magical if the sums of the cells of every row and every column are equal.

The input comes as an array of arrays, containing numbers (number 2D matrix). The input numbers will always be positive.
The output is a Boolean result indicating whether the matrix is magical or not.