City Taxes Task
This task is an extension of Problem 1.
You will receive a city’s name (string), population (number), and treasury (number) as arguments, which you will need to set as properties of an object and return it. In addition to the input parameters, the object must have a property taxRate with an initial value of 10, and three methods for managing the city:

· collectTaxes() - Increase treasury by population * taxRate

· applyGrowth(percentage) - Increase population by given percentage

· applyRecession(percentage) - Decrease treasury by given percentage

Round down the values after each calculation.