Object Factory Task
Create a function that can compose objects by copying functions from a given library of functions.You will receive two parameters – a library of functions as an associative array (object) and an array of orders, represented as objects. You must return a new array – the fulfilled orders. A template is an object that must be copied. The parts array contains the names of required functions as strings.