Construction Crew
Weight is expressed in kilograms, experience in years and levelOfHydrated is in milliliters. If you receive a worker whose dizziness property is set to true it means he needs to intake some water to be able to work correctly. The required amount is 0.1ml per kilogram per year of experience. The required amount must be added to the existing amount (to the levelOfHydrated). Once the water is administered, change the dizziness property to false.

Workers who do not have dizziness should not be modified in any way. Return them as they were.