Subtraction and Pascal or Camel case tasks
1. Subtraction

An HTML page holds two text fields with ids "firstNumber" and "secondNumber". Write a function to subtract the values from these text fields and display the result in the div named "result".
2. Pascal or Camel Case

An HTML file is given and your task is to write a function that takes two string parameters as an input and transforms the first parameter to the type required by the second parameter.

· The first parameter will be the text that you need to modify depending on the second parameter. The words in it will always be separated by space.

· The second parameter will be either "Camel Case" or "Pascal Case". In case of different input, your output should be "Error!