1.Area and Volume Calculator
Write a function that calculates the area and the volume of a figure, which is defined by its coordinates (x, y, z).
You will receive 3 parameters - the functions area and vol and a string, which contains the figures' coordinates.

For more information check the examples.

Output

The output should be returned as an array of objects. Each object has two properties: the figure's area and volume.

[

{ area: ${area1}, volume: ${volume1} },

{ area: ${area2}, volume: ${volume2} },

. . .

]