4. Filter Employees
Write a program that filters the employees of your company. You should print the result in a specific format. You will receive 2 parameters (data, criteria). You should parse the input, find all employees that fulfill the criteria, and print them.

Input

You will receive a string with all the employees, and criteria by which you should sort the employees. If the criteria are "all" print all the employees in the given format.

Output

The output should be printed on the console.