9. Task Manager
Use the index.html and app.js files to solve this problem. You have NO permission to directly change the given HTML code (index.html file).
Your task

Write the missing JavaScript code to make the Task Manager Functionality works as follows:

When the [Add] button is clicked, first you need to validate the inputs. If any of the input fields are empty, the function doesn’t make anything.

After validating the input fields, you need to add the new task (article) in the "Open" section.

The HTML structure looks like this:

The article should have two buttons "Start" and "Delete". Be careful to set the classes for the buttons and the parent-div.

When the [Start] button is clicked, you need to move the Task in the section "In Progress". Be careful with the buttons! The HTML structure looks like this:

When the [Delete] button is clicked, the Task (whole article) should be removed from the HTML.

After clicking the [Finish] button, the Task will be completed, and you should move the article to the section "Complete". The buttons with their parent div-element should be removed.