3. Deck of Cards
Write a function that takes a deck of cards as an array of strings and prints them as a sequence of cards (space separated). Use the solution from the previous task to generate the cards.

Print `Invalid card: ${card}` when an invalid card definition is passed as input.

Input / Output

The function takes an array of strings as a parameter. Print the list of cards as string, separated by space.