6. Dynamic Validation
Write a JS function that dynamically validates an email input field when it is changed. If the input is invalid, apply to it the class "error". Do not validate on every keystroke, as it is annoying for the user, consider only change events.

A valid email will be in format: <name>@<domain>.<extension>

Only lowercase Latin characters are allowed for any of the parts of the email. If the input is valid, clear the style.