1. Getting the information from the form
When you click the ["Submit"] button, the information from the input fields must be added to the ul with the id "in-progress" and then clear input fields. Also the counter with id "progress-count" should be increased by 1. 2. Edit information for posts

When the ["Edit"] button is clicked, the information from the post must be sent to the input’s fields and the record should be deleted from the ul "in-progress". Also the counter with id "progress-count" should be decreased by 1.3. Complete posts

When you click the ["Mark as complete"] button, the record must be deleted from the ul with id

"in-progress" , appended to the ul with the id "finished" and the counter with id "progress-count" should be decreased by 1.4. Clear posts

When you click the ["Clear"] button, the record for all posts must be deleted from the ul with the id "finished".